Mon, Jan 30, 2023, 9:04 PM -05:00

+++ kochan/10_character_strings/exercise_10/07_string_to_float.c
@@ -44,7 +44,7 @@ int main(void)
 
         int digit = text[i] - 48;
 
-        post_decimal = post_decimal * 10 + digit;
+        post_decimal = (float)(post_decimal * 10 + digit);
 
         count++;
     }
